refactor(frontend): queries > hooks, consolidate useFileMutations
